Server and Client Configurations Files
Moving Server and Client configuration specific variables into a [configparser](https://docs.python.org/3.7/library/configparser.html) format for ease of deployment in varying environments.

The goal would be to eventually deploy this app in a containerized environment so we will need to investigate how to handle addressing and/or DNS entries on whatever substrate we deploy this app on.